Add yellow part highlighting to gradeallq page
This adds a hover-over highlighting on the gradeallq page to make
it easier for the teacher to see which score goes with which part of
the question.  Also refactors inline javascript into a separate file.
This fix renames some scoring boxes, so should not be deployed while
teacher may be working.