Bug fix: LTI 1.3, lineitem update seems to confuse Canvas scoremaximum
In copied courses where the points possible had been adjusted in Canvas,
the score didn't seem to be getting scaled properly, sometimes resulting in
scores like 20/10 in Canvas.  My guess is the scoremaximum in the lineitem
update was being used incorrectly by Canvas.  This change skips the
scoremaximum in the lineitem update on launch when a lineitem already exists.