Bug fix: mathjs issue with sci not; jsxgraph issues on numbers that insert as sci not
mathjs treats 1e-6 as euler's number, but 1E-6 as sci not
That's needed for calculated/numfunc questions which may use e
But jsxgraph inserts the number in a string, and JS inserts
scinot as 1e-6.  This rewrites before calling mathjs.
Also, fixes a bug in sci not handling in mathjs.