Adjust LTI 1.3 handling of role=instructor when not on course
If role=instructor, and user is not already a teacher on the course,
it will now check to see if they're a tutor.  If not, it will check
their imathas user rights, and add them as a tutor or teacher as appropriate.

The intent here is to prevent someone without instructor rights on imathas
from getting added as a teacher to a course via their LTI role.  This should
also help since Canvas identifies TAs using an instructor role.